refactor: extract shared llm provider adapter
